moved test to correct location and updated to use latest test data
